+++ src/leoUI.ts
@@ -573,7 +573,7 @@ export class LeoUI extends NullGui {
                         buttons: true
                     }
                 );
-                await g.openUrlHelper(c, p_arg.unl);
+                await g.openUrlOnClick(c, p_arg.unl);
                 void this.launchRefresh();
                 this.loadSearchSettings();
 
@@ -1385,7 +1385,7 @@ export class LeoUI extends NullGui {
             const w_v = c.fileCommands.gnxDict[w_gnx]; // target to change
             if (w_v) {
 
-                if (body !== w_v.b) {
+                if (body !== w_v.b && !g.doHook("bodykey1", { c: c, v: w_v })) {
                     // if different, replace body and set dirty
                     let w_p: Position | undefined;
                     if (c.p.gnx === w_v.gnx) {
@@ -1417,7 +1417,7 @@ export class LeoUI extends NullGui {
                         }
                         // this.clearHeadlineSelection();
                     }
-
+                    g.doHook("bodykey2", { c: c, v: w_v });
                 }
 
             } else {
@@ -1954,6 +1954,13 @@ export class LeoUI extends NullGui {
             this.leoStates.setSelectedNodeFlags(p_node);
         }
 
+        // Get 'c' before timeout for proper c that was refreshed.
+        const c = g.app.windowList[this.frameIndex].c;
+
+        // Do in a timeout to let rest of tree refresh.
+        setTimeout(() => {
+            g.doHook('after-redraw-outline', { c: c });
+        }, 0);
     }
 
     /**
@@ -2880,6 +2887,8 @@ export class LeoUI extends NullGui {
 
         const c = g.app.windowList[this.frameIndex].c;
 
+        g.doHook("headclick1", { c: c, p: p_node, v: p_node });
+
         // * check if used via context menu's "open-aside" on an unselected node: check if p_node is currently selected, if not select it
         if (
             p_aside &&
@@ -2901,6 +2910,7 @@ export class LeoUI extends NullGui {
         // (other tree nodes with same gnx may have different syntax language coloring because of parents lineage)
         if (p_node.__eq__(this.lastSelectedNode)) {
             this._locateOpenedBody(p_node.gnx); // LOCATE NEW GNX
+            g.doHook("headclick2", { c: c, p: p_node, v: p_node });
             return this.showBody(!!p_aside, w_showBodyKeepFocus).catch((p_error) => {
                 return Promise.resolve(); // intercept cancellation as success: next one is going to replace anyways.
             });
@@ -2925,12 +2935,13 @@ export class LeoUI extends NullGui {
                         // states: false,
                     }
                 );
+                g.doHook("headclick2", { c: c, p: p_node, v: p_node });
                 return this._launchRefresh();
             }
             this._refreshType.states = true;
             this.getStates();
         }
-
+        g.doHook("headclick2", { c: c, p: p_node, v: p_node });
         // * Apply the node to the body text without waiting for the selection promise to resolve
         return this._tryApplyNodeToBody(p_node, !!p_aside, w_showBodyKeepFocus);
 
@@ -3324,13 +3335,23 @@ export class LeoUI extends NullGui {
             this._hibResolve = p_resolve;
             // onDidHide handles CANCEL AND ACCEPT AND INTERCEPT !
             this._hibDisposables.push(hib.onDidHide(() => {
+                let changed = false;
+                const c = g.app.windowList[this.frameIndex].c;
                 if (this._hib) {
-                    this._hibLastValue = this._hib.value; // * FORCE VALUE EVEN WHEN CANCELLING LIKE IN ORIGINAL LEO !
+                    changed = this._hib.value !== p_options.value;
+                    if (g.doHook("headkey1", { c: c, p: c.p, changed: changed })) {
+                        // The hook claims to have handled the event, so 'cancel'.
+                        this._hibLastValue = undefined;
+                    } else {
+                        // ok
+                        this._hibLastValue = this._hib.value; // * FORCE VALUE EVEN WHEN CANCELLING LIKE IN ORIGINAL LEO !
+                    }
                 }
                 this.leoStates.leoEditHeadline = false;
                 if (this._hibResolve) {
                     // RESOLVE whatever value was set otherwise undefined will mean 'canceled'.
                     this._hibResolve(this._hibLastValue);
+                    g.doHook("headkey2", { c: c, p: c.p, changed: changed });
                     // Dispose of everything disposable with the edit headline process.
                     for (const disp of this._hibDisposables) {
                         disp.dispose();
@@ -3426,7 +3447,7 @@ export class LeoUI extends NullGui {
      * * Asks for a headline label to be entered and creates (inserts) a new node under the current, or specified, node
      * @param p_node specified under which node to insert, or leave undefined to use whichever is currently selected
      * @param p_fromOutline Signifies that the focus was, and should be brought back to, the outline
-     * @param p_interrupt Signifies the insert action is actually interrupting itself (e.g. rapid CTRL+I actions by the user)
+     * @param p_asChild Insert as child instead of as sibling
      * @returns Thenable that resolves when done
      */
     public async insertNode(p_node: Position | undefined, p_fromOutline: boolean, p_asChild: boolean): Promise<unknown> {